Add files via upload
Created a new branch. Edited the files so that they can be compiled together without build failure. Parser was dropped, so we only need "compiler.h" and it holds most everything now. Some problems with parser/code generator still. As far as I can tell, the scanner works fine. Didn't get to test the p-machine. Focus on the parser because that alone is due.